Move Persona into persona.ts
